test(e2e): rename server constants to puppeteer constants
