+++ test/e2e/ClientOptions.test.js
@@ -7,7 +7,7 @@ const testServer = require('../helpers/test-server');
 const config = require('../fixtures/client-config/webpack.config');
 const runBrowser = require('../helpers/run-browser');
 const [port1, port2, port3] = require('../ports-map').ClientOptions;
-const { beforeBrowserCloseDelay } = require('../helpers/server-constants');
+const { beforeBrowserCloseDelay } = require('../helpers/puppeteer-constants');
 
 describe('Client code', () => {
   function startProxy(port) {
