+++ test/e2e/Progress.test.js
@@ -11,7 +11,7 @@ const port = require('../ports-map').Progress;
 const {
   reloadReadyDelay,
   completeReloadDelay,
-} = require('../helpers/server-constants');
+} = require('../helpers/puppeteer-constants');
 
 const cssFilePath = resolve(__dirname, '../fixtures/reload-config/main.css');
